Filter out non-process facets, and store as array
We add a step to the pre-processing to filter out irrelevant
facets, based on the criteria that all 'relevant' facets,
i.e. those that have a considerable energy impact, start
with F28. Moreover, we now also store them as an array in
the recipes json, rather than a 569JVXZghimsseparated string