Changes to dried fruits and juices
Previously, dried fruits and juices were immediately broken down in the SFA recipe to their RPC with a reverse yield, i.e. orange juice --> oranges with a reverse yield of 5. This caused problems as the allocation factor was not taken into account, nor special transport / packaging considerations for certain processes as these are tied to the EFSA-recipes. Therefore, changed these kinds of codes to their actual EFSA-equivalent (tying orange juice to the FoodEx2 code for orange juice, instead of oranges).